Try to use both ways how to uninstall server
Since current master (future 4.5) has trouble uninstalling last
master server we need to try both approaches.

This is necessary for the cases in which a job failed before
attempting an uninstall in previous scheduled job. The new job does
not have any indication which version of server is installed,
therefore both methods are tried.